Moved the end date search filter into the search view.
