+++ aliss/tests/views/test_search_view.py
@@ -3,6 +3,8 @@ from django.test import Client
 from django.urls import reverse
 from aliss.tests.fixtures import Fixtures
 from aliss.models import *
+from datetime import datetime, timedelta
+import pytz
 
 class SearchViewTestCase(TestCase):
     fixtures = ['service_areas.json', 'g2_postcodes.json']
@@ -290,6 +292,30 @@ class SearchViewTestCase(TestCase):
         response = self.client.get('/search/?postcode=AB20')
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "<h1>Sorry, AB20 doesn't appear to be a valid postcode.</h1>", html=True)
+
+    def test_filter_by_end_date_one_ended_service(self):
+        utc = pytz.UTC
+        current_date = datetime.now()
+        current_date = utc.localize(current_date)
+        one_week_ago = (current_date - timedelta(weeks=1))
+        response = self.client.get('/search/?postcode=G2+4AA')
+        self.assertContains(response, "My Testing Service")
+        self.s2.end_date = one_week_ago
+        self.s2.save()
+        response = self.client.get('/search/?postcode=G2+4AA')
+        self.assertNotContains(response, "My Testing Service")
+
+    def test_filter_by_end_date_one_service_ended_in_future(self):
+        utc = pytz.UTC
+        current_date = datetime.now()
+        current_date = utc.localize(current_date)
+        one_week_in_future = (current_date + timedelta(weeks=1))
+        response = self.client.get('/search/?postcode=G2+4AA')
+        self.assertContains(response, "My Testing Service")
+        self.s2.end_date = one_week_in_future
+        self.s2.save()
+        response = self.client.get('/search/?postcode=G2+4AA')
+        self.assertContains(response, "My Testing Service")
 
     def tearDown(self):
         Fixtures.organisation_teardown()
