+++ aliss/views/search.py
@@ -8,6 +8,9 @@ from django.http import HttpResponseRedirect
 from elasticsearch_dsl import Search
 from elasticsearch_dsl.connections import connections
 
+from datetime import datetime, timedelta
+import pytz
+
 from aliss.paginators import ESPaginator
 from aliss.forms import SearchForm
 from aliss.models import Postcode, Service, Category
@@ -95,6 +98,13 @@ class SearchView(MultipleObjectMixin, TemplateView):
         return queryset
 
     def filter_queryset(self, queryset):
+        '''
+        Add the current date to allow comparison with end_date
+        '''
+        utc = pytz.UTC
+        current_date = datetime.now()
+        current_date = utc.localize(current_date)
+
         if self.category:
             queryset = filter_by_category(queryset, self.category)
         if self.location_type:
@@ -111,7 +121,8 @@ class SearchView(MultipleObjectMixin, TemplateView):
         else:
             results = postcode_order(queryset, self.postcode)
         self.distance_scores = self.check_distance_within_radius(results["distance_scores"], self.radius)
-        return Service.objects.filter(id__in=results["ids"]).order_by(results["order"])
+        services = Service.objects.filter(id__in=results["ids"]).order_by(results["order"]).exclude(end_date__lt=current_date)
+        return services
 
     def assign_legacy_postcode(self, location, legacy_locations_dict):
         postcode = Postcode.objects.get(postcode = legacy_locations_dict.get(str(location)))
